+++ lib/src/parser/models/config/ios.g.dart
@@ -6,9 +6,18 @@ part of 'ios.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-IOS _$IOSFromJson(Map json) => IOS(
-      buildSettings: (json['buildSettings'] as Map?)?.map(
-            (k, e) => MapEntry(k as String, e),
-          ) ??
-          {},
-    );
+IOS _$IOSFromJson(Map json) {
+  $checkKeys(json, disallowNullValues: const ['includes']);
+  return IOS(
+    buildSettings:
+        (json['buildSettings'] as Map?)?.map(
+          (k, e) => MapEntry(k as String, e),
+        ) ??
+        {},
+    includes:
+        (json['includes'] as List<dynamic>?)
+            ?.map((e) => Include.fromJson(Map<String, dynamic>.from(e as Map)))
+            .toList() ??
+        [],
+  );
+}