+++ lib/src/parser/models/flavors/darwin.dart
@@ -25,6 +25,7 @@
 
 import 'package:flutter_flavorizr/src/parser/mixins/build_settings_mixin.dart';
 import 'package:flutter_flavorizr/src/parser/models/flavors/commons/os.dart';
+import 'package:flutter_flavorizr/src/parser/models/flavors/darwin/include.dart';
 import 'package:flutter_flavorizr/src/parser/models/flavors/darwin/variable.dart';
 import 'package:flutter_flavorizr/src/parser/models/flavors/google/firebase/firebase.dart';
 import 'package:json_annotation/json_annotation.dart';
@@ -39,9 +40,13 @@ class Darwin extends OS with BuildSettingsMixin {
   @JsonKey(disallowNullValue: true, defaultValue: {})
   final Map<String, Variable> variables;
 
+  @JsonKey(disallowNullValue: true, defaultValue: [])
+  final List<Include> includes;
+
   Darwin({
     required this.bundleId,
     this.variables = const {},
+    this.includes = const [],
     Map<String, dynamic> buildSettings = const {},
     super.generateDummyAssets,
     super.firebase,