+++ test/processors/macos/xcconfig_processor_test.dart
@@ -55,8 +55,118 @@ void main() {
             'test_resources/macos/xcconfig_processor_test/matcher_without_variables.xcconfig')
         .readAsStringSync();
 
+    final flavorName = flavorizr.flavors.keys.first;
+    final flavor = flavorizr.flavors[flavorName]!;
+
     MacOSXCConfigProcessor processor = MacOSXCConfigProcessor(
+      flavorName,
+      flavor,
       Target.debug,
+      config: flavorizr,
+      logger: logger,
+    );
+    String actual = processor.execute();
+
+    actual = TestUtils.stripEndOfLines(actual);
+    matcher = TestUtils.stripEndOfLines(matcher);
+
+    expect(actual, matcher);
+  });
+
+  test('Test MacOSXCConfigProcessor with global includes', () {
+    logger = Logger(level: Level.quiet);
+    Parser parser = const Parser(
+      pubspecPath:
+          'test_resources/macos/xcconfig_processor_test/pubspec_with_global_includes',
+      flavorizrPath: 'test_resources/non_existent',
+    );
+    try {
+      flavorizr = parser.parse();
+    } catch (e) {
+      fail(e.toString());
+    }
+
+    String matcher = File(
+            'test_resources/macos/xcconfig_processor_test/matcher_with_global_includes.xcconfig')
+        .readAsStringSync();
+
+    final flavorName = flavorizr.flavors.keys.first;
+    final flavor = flavorizr.flavors[flavorName]!;
+
+    MacOSXCConfigProcessor processor = MacOSXCConfigProcessor(
+      flavorName,
+      flavor,
+      Target.debug,
+      config: flavorizr,
+      logger: logger,
+    );
+    String actual = processor.execute();
+
+    actual = TestUtils.stripEndOfLines(actual);
+    matcher = TestUtils.stripEndOfLines(matcher);
+
+    expect(actual, matcher);
+  });
+
+  test('Test MacOSXCConfigProcessor with flavor includes', () {
+    logger = Logger(level: Level.quiet);
+    Parser parser = const Parser(
+      pubspecPath:
+          'test_resources/macos/xcconfig_processor_test/pubspec_with_flavor_includes',
+      flavorizrPath: 'test_resources/non_existent',
+    );
+    try {
+      flavorizr = parser.parse();
+    } catch (e) {
+      fail(e.toString());
+    }
+
+    String matcher = File(
+            'test_resources/macos/xcconfig_processor_test/matcher_with_flavor_includes.xcconfig')
+        .readAsStringSync();
+
+    final flavorName = flavorizr.flavors.keys.first;
+    final flavor = flavorizr.flavors[flavorName]!;
+
+    MacOSXCConfigProcessor processor = MacOSXCConfigProcessor(
+      flavorName,
+      flavor,
+      Target.debug,
+      config: flavorizr,
+      logger: logger,
+    );
+    String actual = processor.execute();
+
+    actual = TestUtils.stripEndOfLines(actual);
+    matcher = TestUtils.stripEndOfLines(matcher);
+
+    expect(actual, matcher);
+  });
+
+  test('Test MacOSXCConfigProcessor with includes and target', () {
+    logger = Logger(level: Level.quiet);
+    Parser parser = const Parser(
+      pubspecPath:
+          'test_resources/macos/xcconfig_processor_test/pubspec_with_includes_and_target',
+      flavorizrPath: 'test_resources/non_existent',
+    );
+    try {
+      flavorizr = parser.parse();
+    } catch (e) {
+      fail(e.toString());
+    }
+
+    String matcher = File(
+            'test_resources/macos/xcconfig_processor_test/matcher_with_includes_and_target.xcconfig')
+        .readAsStringSync();
+
+    final flavorName = flavorizr.flavors.keys.first;
+    final flavor = flavorizr.flavors[flavorName]!;
+
+    MacOSXCConfigProcessor processor = MacOSXCConfigProcessor(
+      flavorName,
+      flavor,
+      Target.release,
       config: flavorizr,
       logger: logger,
     );